feat: Added scaffolding for change handle unit tests
