+++ pallets/handles/src/tests/mod.rs
@@ -1,6 +1,7 @@
 pub mod mock;
 
 mod handle_creation_tests;
+mod handle_change_tests;
 mod handle_retirements_tests;
 mod handles_replay_attack_test;
 mod signed_extension_tests;
